backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=3.10,<4.0
djangorestframework-simplejwt>=5.0,<6.0
pytest>=6.2,<7.0
pytest-django>=4.4,<5.0
gunicorn>=20.1,<21.0
whitenoise>=6.0,<7.0
dj-database-url>=1.0,<2.0